Refactor SAST get results logic from FlowService into ResultsService
