+++ src/test/java/com/checkmarx/flow/cucumber/component/batch/BatchComponentSteps.java
@@ -7,6 +7,7 @@ import com.checkmarx.flow.cucumber.common.utils.TestUtils;
 import com.checkmarx.flow.cucumber.component.scan.ScanFixture;
 import com.checkmarx.flow.service.FlowService;
 import com.checkmarx.flow.service.HelperService;
+import com.checkmarx.flow.service.ResultsService;
 import com.checkmarx.flow.service.SastScannerService;
 import com.checkmarx.sdk.config.CxProperties;
 import com.checkmarx.sdk.exception.CheckmarxException;
@@ -38,13 +39,14 @@ public class BatchComponentSteps {
     private final HelperService helperService;
     private final CxClient cxClient;
     private final List<ThreadPoolTaskExecutor> executors;
+    private final ResultsService resultsService;
     private CxFlowRunner cxFlowRunner;
     private String projectName;
     private String teamName;
     private SastScannerService sastScannerService;
 
     public BatchComponentSteps(FlowProperties flowProperties, CxProperties cxProperties, JiraProperties jiraProperties, GitHubProperties gitHubProperties, GitLabProperties gitLabProperties,
-                               ADOProperties adoProperties, FlowService flowService, HelperService helperService, CxClient cxClient, List<ThreadPoolTaskExecutor> executors, SastScannerService sastScannerService) {
+                               ADOProperties adoProperties, FlowService flowService, HelperService helperService, CxClient cxClient, List<ThreadPoolTaskExecutor> executors, SastScannerService sastScannerService, ResultsService resultsService) {
         this.flowProperties = flowProperties;
         this.cxProperties = cxProperties;
         this.jiraProperties = jiraProperties;
@@ -55,6 +57,7 @@ public class BatchComponentSteps {
         this.helperService = helperService;
         this.cxClient = cxClient;
         this.executors = executors;
+        this.resultsService = resultsService;
     }
 
     @Given("SAST client is mocked - to allow tests to pass without active SAST environment")
@@ -62,7 +65,7 @@ public class BatchComponentSteps {
         when(cxClient.getTeamId(anyString())).thenReturn(ScanFixture.TEAM_ID);
         when(cxClient.getProjects(anyString())).thenReturn(ScanFixture.getProjects());
         when(cxClient.getReportContentByScanId(ScanFixture.SCAN_ID, ScanFixture.getScanFilters())).thenReturn(ScanFixture.getScanResults());
-        cxFlowRunner = new CxFlowRunner(flowProperties, cxProperties, jiraProperties, gitHubProperties, gitLabProperties, adoProperties, helperService, flowService, sastScannerService, executors);
+        cxFlowRunner = new CxFlowRunner(flowProperties, cxProperties, jiraProperties, gitHubProperties, gitLabProperties, adoProperties, helperService, flowService, sastScannerService, executors, resultsService);
     }
 
     @Given("project is provided: {string} and team: {string}")
